pyproject.toml
[tool.poetry]
name = "gensor"
version = "0.0.4"
description = "Library for handling groundwater sensor data."
authors = ["Mateusz Zawadzki <zawadzkimat@outlook.com>"]
repository = "https://github.com/zawadzkim/gensor"
documentation = "https://zawadzkim.github.io/gensor/"
readme = "README.md"
packages = [
  {include = "gensor"},
  {include = "py.typed"}
]

[tool.poetry.dependencies]
python = ">=3.11"
pydantic = "^2.8.2"
pandera = "^0.20.3"
chardet = "^5.2.0"
matplotlib = "^3.9.2"
scipy = "^1.14.1"
numpy = "^2.1.0"
scikit-learn = "^1.5.1"
sqlalchemy = "^2.0.32"
pandas = "^2.2.2"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
deptry = "^0.16.2"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"
jupyter = "^1.1.1"
pandas-stubs = "^2.2.2.240807"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.34"
mkdocstrings-python = "^1.11.1"
mkdocs-autorefs = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["gensor"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"
ignore_missing_imports = "True"
disable_error_code = ["union-attr", "no-untyped-def", "return-value", "arg-type", "assignment", "call-arg"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 88
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["gensor"]
